requirement.txt
absl-py==2.1.0
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==22.3.0
bleach==6.2.0
blis==1.0.2
bs4==0.0.2
catalogue==2.0.10
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.10
Cython==0.29.37
cython-lint==0.16.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.4
djangorestframework==3.15.2
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
flake8==5.0.4
flatbuffers==24.3.25
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
html2text==2024.2.26
huggingface-hub==0.30.2
hypothesis==6.122.3
identify==2.6.3
idna==3.10
iniconfig==2.0.0
ipython==8.12.3
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.7.0
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-datasets==0.2.0
ml-dtypes==0.4.1
mock==2.0.0
msvc_runtime==14.42.34433
murmurhash==1.0.11
mypy==1.5.1
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nltk==3.9.1
nodeenv==1.9.1
numpy==2.0.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pbr==6.1.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==5.29.1
pure_eval==0.2.3
pycodestyle==2.9.1
pydantic==2.10.3
pydantic_core==2.27.1
pyflakes==2.5.0
Pygments==2.18.0
pypdf==5.1.0
pytest==8.3.4
pytest-timeout==1.4.2
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
shellingham==1.5.4
six==1.17.0
smart-open==7.0.5
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.0
stack-data==0.6.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
thinc==8.3.2
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenize_rt==6.1.0
tokenizers==0.21.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
types-mock==5.1.0.20240425
types-requests==2.32.0.20241016
types-setuptools==75.6.0.20241126
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.0
yarg==0.1.9
